Converted floats to doubles

+++ include/layer.h
@@ -6,9 +6,9 @@ class Layer{
         Layer(int size, int no_inputs, bool last_layer);
 
         std::vector<Neuron> neurons_;
-        std::vector<float> inputs_;
-        std::vector<float> outputs_;
+        std::vector<double> inputs_;
+        std::vector<double> outputs_;
 
-        std::vector<float> calculate_layer_outputs();
+        std::vector<double> calculate_layer_outputs();
 
 };
