+++ src/neuron.cpp
@@ -3,14 +3,14 @@
 Neuron::Neuron(int no_inputs, bool last_layer){
     // Initialisation
     for(int i = 0; i < no_inputs; i++){
-        float random_weight = ((float) rand() / RAND_MAX) * 2.0f - 1.0f;
+        double random_weight = ((double) rand() / RAND_MAX) * 2.0f - 1.0f;
         weights_.push_back(random_weight);
     }
-    bias_ = ((float) rand() / RAND_MAX) * 2.0f - 1.0f;
+    bias_ = ((double) rand() / RAND_MAX) * 2.0f - 1.0f;
 };
 
-float Neuron::calculate_neuron_output(){
-    float sum = 0;
+double Neuron::calculate_neuron_output(){
+    double sum = 0;
     for(int i = 0; i < inputs_.size(); i++){
         sum += inputs_[i] * weights_[i];
     }
